Optimize

+++ Sources/LiveKitComponents/UI/Visualizer/Visualizer.swift
@@ -169,14 +169,14 @@ public struct BarAudioVisualizer: View {
 
     public var body: some View {
         GeometryReader { geometry in
-            let barMinHeight = (CGFloat(geometry.size.width) - CGFloat(geometry.size.width * barSpacingFactor) * CGFloat(barCount + 2)) / CGFloat(barCount)
+            let barMinHeight = (geometry.size.width - geometry.size.width * barSpacingFactor * CGFloat(barCount + 2)) / CGFloat(barCount)
             HStack(alignment: .center, spacing: geometry.size.width * barSpacingFactor) {
                 ForEach(0 ..< audioProcessor.bands.count, id: \.self) { index in
                     VStack {
                         Spacer()
                         RoundedRectangle(cornerRadius: barMinHeight)
-                            .fill(barColor.opacity(Double(1.0 - barMinOpacity) * Double(audioProcessor.bands[index]) + barMinOpacity))
-                            .frame(height: ((geometry.size.height - barMinHeight) * CGFloat(audioProcessor.bands[index])) + barMinHeight)
+                            .fill(barColor.opacity((1.0 - barMinOpacity) * Double(audioProcessor.bands[index]) + barMinOpacity))
+                            .frame(height: (geometry.size.height - barMinHeight) * CGFloat(audioProcessor.bands[index]) + barMinHeight)
                         Spacer()
                     }
                 }
